Update cps3_to_new_names3
in above program we have to upload file on s3 and we already create one lambda where we provide trigger on create event as we upload the file in s3 lambda will trigger. In above lambda code we are copying the file name which is created or modified recently in bucket input location and rename that file and save in different output location of same bucket.